Revert "Move versioneer into _setup"
This reverts commit 6ef1827872275ffd9ea2544b6028a4a3b4635a66.
